refactor(core): rename `composiable` to `composables`
