+++ packages/core/src/index.ts
@@ -1,7 +1,7 @@
-export * from './composable/useForm';
-export * from './composable/useField';
-export * from './composable/useFieldArray';
-export { useFormContext } from './composable/useFormContext';
+export * from './composables/useForm';
+export * from './composables/useField';
+export * from './composables/useFieldArray';
+export { useFormContext } from './composables/useFormContext';
 export { default as set } from './utils/set';
 export { default as get } from './utils/get';
 export * from './types';
